+++ services/web/src/components/pages/Home/components/DefaultHomePage/StatusBar.tsx
@@ -7,7 +7,7 @@ import {
 
 import useMediaQuery from '@material-ui/core/useMediaQuery';
 
-import styles from '../../../../../../styles/pages/statusBar.module.css';
+import styles from '../../../../../styles/pages/statusBar.module.css';
 
 const { Step } = Steps;
